+++ packages/debug_extension/lib/metrics.dart
@@ -1,5 +1,5 @@
-export 'src/metrics/metrics_plugin.dart';
-export 'src/metrics/metrics_plugin_platform_interface.dart';
-export 'src/metrics/metrics_plugin_method_channel.dart';
+export 'src/plugin/debug_extension_plugin.dart';
+export 'src/plugin/debug_extension_plugin_platform_interface.dart';
+export 'src/plugin/debug_extension_plugin_method_channel.dart';
 export 'src/metrics/metrics_data.dart';
 export 'src/metrics/metrics_collector.dart';